@@ -5,7 +5,7 @@ buildscript {
 
     dependencies {
         //find the latest io.fabric.tools:gradle here: https://docs.fabric.io/android/changelog.html#fabric-gradle-plugin
-        classpath 'io.fabric.tools:gradle:1.25.4'
+        classpath 'io.fabric.tools:gradle:1.28.0'
     }
 }
 apply plugin: 'com.android.application'
@@ -27,7 +27,7 @@ otherPrivateProperties.load(new FileInputStream(otherPrivatePropertiesFile))
 def minSdk = hasProperty('minSdk') ? minSdk : 15
 
 ext {
-    supportLibVersion = '27.1.1'
+    androidxVersion = '1.0.0'
 }
 
 android {
@@ -39,7 +39,7 @@ android {
             storePassword keystoreProperties['storePassword']
         }
     }
-    compileSdkVersion 26
+    compileSdkVersion 28
     buildToolsVersion '28.0.3'
     defaultConfig {
         vectorDrawables.useSupportLibrary = true
@@ -47,8 +47,8 @@ android {
         applicationId 'com.myhitchhikingspots'
         minSdkVersion minSdk
         targetSdkVersion 26
-        versionCode 26
-        versionName "2.2.4"
+        versionCode 27
+        versionName "2.2.5"
         // Stops the Gradle plugin’s automatic rasterization of vectors
         generatedDensities = []
         // Enabling multidex support.
@@ -81,13 +81,16 @@ android {
 repositories {
     mavenCentral()
     maven { url 'https://maven.fabric.io/public' }
+    maven { url "https://maven.google.com/" }
 }
 
 dependencies {
     //Begin: Dependencies for MapBox
     // add the Mapbox SDK dependency below
-    implementation('com.mapbox.mapboxsdk:mapbox-android-sdk:6.4.0')
-    implementation 'com.mapbox.mapboxsdk:mapbox-android-plugin-locationlayer:0.7.2'
+    implementation('com.mapbox.mapboxsdk:mapbox-android-sdk:7.3.2')
+
+    //Mapbox plugins
+    implementation 'com.mapbox.mapboxsdk:mapbox-android-plugin-localization-v7:0.9.0'
 
     //End: Dependencies for MapBox
     implementation 'cz.msebera.android:httpclient:4.4.1.1'
@@ -97,14 +100,17 @@ dependencies {
     implementation 'org.greenrobot:greendao:3.2.0'
     implementation 'com.google.android.gms:play-services-location:16.0.0'
     implementation 'joda-time:joda-time:2.9.2'
-    implementation 'com.android.support:multidex:1.0.3'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
+    implementation 'androidx.multidex:multidex:2.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     implementation 'com.github.florent37:viewtooltip:1.1.5'
-    implementation "com.android.support:appcompat-v7:${supportLibVersion}"
-    implementation "com.android.support:recyclerview-v7:${supportLibVersion}"
-    implementation "com.android.support:design:${supportLibVersion}"
+
+    implementation "androidx.appcompat:appcompat:${androidxVersion}"
+    implementation "androidx.recyclerview:recyclerview:${androidxVersion}"
+    implementation "com.google.android.material:material:${androidxVersion}"
+    implementation "androidx.cardview:cardview:${androidxVersion}"
 
     implementation "com.squareup.picasso:picasso:2.5.2"
-    implementation "com.android.support:cardview-v7:${supportLibVersion}"
-
+    testImplementation "junit:junit:4.13-beta-2"
+    testImplementation "com.google.truth:truth:0.44"
+    implementation "com.google.firebase:firebase-core:16.0.8"
 }
