@@ -21,10 +21,10 @@ import java.util.ArrayList;
 // THIS CODE IS GENERATED BY greenDAO, DO NOT EDIT.
 
 /**
- * Master of DAO (schema version 5): knows all DAOs.
+ * Master of DAO (schema version 6): knows all DAOs.
  */
 public class DaoMaster extends AbstractDaoMaster {
-    public static final int SCHEMA_VERSION = 5;
+    public static final int SCHEMA_VERSION = 6;
 
     /**
      * Creates underlying database table using DAOs.
@@ -147,7 +147,10 @@ public class DaoMaster extends AbstractDaoMaster {
                         "ALTER TABLE " + SpotDao.TABLENAME + " ADD '" + SpotDao.Properties.IsHitchhikingSpot.columnName + "' " + SpotDao.Properties.IsHitchhikingSpot.type.getSimpleName() + " DEFAULT " + Constants.ISHITCHHIKINGSPOT_DEFAULT_VALUE + ";",
                         "UPDATE " + SpotDao.TABLENAME +
                                 " SET " + SpotDao.Properties.IsHitchhikingSpot.columnName + " = '0' " +
-                                " WHERE " + SpotDao.Properties.IsDestination.columnName + " = '1'"
+                                " WHERE " + SpotDao.Properties.IsDestination.columnName + " = '1';"
+                };
+                String[] changesRequeriedByVersion6 = new String[]{
+                        "ALTER TABLE " + SpotDao.TABLENAME + " ADD '" + SpotDao.Properties.IsNotHitchhikedFromHere.columnName + "' " + SpotDao.Properties.IsNotHitchhikedFromHere.type.getSimpleName() + " DEFAULT 0;"
                 };
 
                 ArrayList<String[]> versionsUpdate = new ArrayList<>();
@@ -155,28 +158,18 @@ public class DaoMaster extends AbstractDaoMaster {
                 versionsUpdate.add(changesRequeriedByVersion3);
                 versionsUpdate.add(changesRequeriedByVersion4);
                 versionsUpdate.add(changesRequeriedByVersion5);
+                versionsUpdate.add(changesRequeriedByVersion6);
                 //WARN: FOR NEW COLUMN ADDED, REMEMBER TO UPDATE DatabaseImporter IF A DEFAULT VALUE SHOULD BE SET!
 
-                String changesRequiredByNewVersion = "";
 
                 for (int i = oldVersion; i < newVersion && i <= versionsUpdate.size(); i++) {
                     //Subtract 1 from oldVersion because the versions started from version 1 but the arrays start from 0
-                    for (String s : versionsUpdate.get(i - 1))
-                        changesRequiredByNewVersion += s;
+                    for (String s : versionsUpdate.get(i - 1)) {
+                        Crashlytics.log("Will try to execute changesRequiredByNewVersion:\n" + s);
+                        db.execSQL(s);
+                        Crashlytics.log("Changes executed successfully");
+                    }
                 }
-
-                //1,2 - 0
-                //1,3 - 0,1
-                //2,3 - 1
-                //2,4 - 1,2
-
-                if (!changesRequiredByNewVersion.isEmpty()) {
-                    Crashlytics.log(Log.INFO, "greenDAO", "Will try to execute changesRequiredByNewVersion:\n" + changesRequiredByNewVersion);
-                    db.execSQL(changesRequiredByNewVersion);
-                    Crashlytics.log(Log.INFO, "greenDAO", "Changes executed successfully");
-                } else
-                    Crashlytics.log(Log.WARN, "greenDAO", "Nothing was changed on the database schema from version " + oldVersion + " to version " + newVersion + "?");
-
 
                 /*dropAllTables(db, true);
                         onCreate(db);
