@@ -1,6 +1,8 @@
 package com.myhitchhikingspots.model;
 
 import org.greenrobot.greendao.annotation.*;
+import org.joda.time.DateTime;
+import org.joda.time.DateTimeZone;
 
 // THIS CODE IS GENERATED BY greenDAO, DO NOT EDIT. Enable "keep" sections if you want to edit.
 
@@ -24,7 +26,7 @@ public class Spot implements java.io.Serializable {
     private Boolean IsReverseGeocoded;
     private String Note;
     private String Description;
-    private java.util.Date StartDateTime;
+    private Long StartDateTime;
     private Integer WaitingTime;
     private Integer Hitchability;
     private Integer AttemptResult;
@@ -35,6 +37,7 @@ public class Spot implements java.io.Serializable {
     private Float Accuracy;
     private Boolean IsPartOfARoute;
     private Boolean IsHitchhikingSpot;
+    private Boolean IsNotHitchhikedFromHere;
 
     @Generated
     public Spot() {
@@ -45,7 +48,7 @@ public class Spot implements java.io.Serializable {
     }
 
     @Generated
-    public Spot(Long id, String Name, String Street, String Zip, String City, String State, String Country, Double Longitude, Double Latitude, Boolean GpsResolved, Boolean IsReverseGeocoded, String Note, String Description, java.util.Date StartDateTime, Integer WaitingTime, Integer Hitchability, Integer AttemptResult, Boolean IsWaitingForARide, Boolean IsDestination, String CountryCode, Boolean HasAccuracy, Float Accuracy, Boolean IsPartOfARoute, Boolean IsHitchhikingSpot) {
+    public Spot(Long id, String Name, String Street, String Zip, String City, String State, String Country, Double Longitude, Double Latitude, Boolean GpsResolved, Boolean IsReverseGeocoded, String Note, String Description, Long StartDateTime, Integer WaitingTime, Integer Hitchability, Integer AttemptResult, Boolean IsWaitingForARide, Boolean IsDestination, String CountryCode, Boolean HasAccuracy, Float Accuracy, Boolean IsPartOfARoute, Boolean IsHitchhikingSpot, Boolean IsNotHitchhikedFromHere) {
         this.id = id;
         this.Name = Name;
         this.Street = Street;
@@ -70,6 +73,7 @@ public class Spot implements java.io.Serializable {
         this.Accuracy = Accuracy;
         this.IsPartOfARoute = IsPartOfARoute;
         this.IsHitchhikingSpot = IsHitchhikingSpot;
+        this.IsNotHitchhikedFromHere = IsNotHitchhikedFromHere;
     }
 
     public Long getId() {
@@ -176,12 +180,32 @@ public class Spot implements java.io.Serializable {
         this.Description = Description;
     }
 
-    public java.util.Date getStartDateTime() {
+    public Long getStartDateTimeMillis() {
         return StartDateTime;
     }
 
-    public void setStartDateTime(java.util.Date StartDateTime) {
-        this.StartDateTime = StartDateTime;
+    public void setStartDateTimeMillis(Long StartDateTimeMillis) {
+        this.StartDateTime = StartDateTimeMillis;
+    }
+
+    /**
+     * Gets the start date and time when the hitchhiker began hitchhiking at this spot.
+     * To avoid confusion with datetime, this datetime is always in UTC.
+     * E.g. If when the hitchhiker began hitchhiking at '1970-01-01 10:00 BRL' (his local time), then StartDateTime here is '1970-01-01 10:00 UTC'.
+     * **/
+    public DateTime getStartDateTime() {
+        if (StartDateTime != null)
+            return new DateTime(getStartDateTimeMillis(), DateTimeZone.UTC);
+        return null;
+    }
+
+    /**
+     * @param StartDateTime The start date and time when the hitchhiker began hitchhiking at this spot.
+     *                      To avoid confusion with datetime, this datetime is always in UTC.
+     *                      E.g. If the local time for the hitchhiker is '1970-01-01 10:00 BRL', then StartDateTime here is '1970-01-01 10:00 UTC'.
+     * **/
+    public void setStartDateTime(DateTime StartDateTime) {
+        setStartDateTimeMillis(StartDateTime.getMillis());
     }
 
     public Integer getWaitingTime() {
@@ -262,6 +286,14 @@ public class Spot implements java.io.Serializable {
 
     public void setIsHitchhikingSpot(Boolean IsHitchhikingSpot) {
         this.IsHitchhikingSpot = IsHitchhikingSpot;
+    }
+
+    public Boolean getIsNotHitchhikedFromHere() {
+        return IsNotHitchhikedFromHere;
+    }
+
+    public void setIsNotHitchhikedFromHere(Boolean IsNotHitchhikedFromHere) {
+        this.IsNotHitchhikedFromHere = IsNotHitchhikedFromHere;
     }
 
 }
