@@ -35,7 +35,7 @@ public class SpotDao extends AbstractDao<Spot, Long> {
         public final static Property IsReverseGeocoded = new Property(10, Boolean.class, "IsReverseGeocoded", false, "IS_REVERSE_GEOCODED");
         public final static Property Note = new Property(11, String.class, "Note", false, "NOTE");
         public final static Property Description = new Property(12, String.class, "Description", false, "DESCRIPTION");
-        public final static Property StartDateTime = new Property(13, java.util.Date.class, "StartDateTime", false, "START_DATE_TIME");
+        public final static Property StartDateTime = new Property(13, Long.class, "StartDateTime", false, "START_DATE_TIME");
         public final static Property WaitingTime = new Property(14, Integer.class, "WaitingTime", false, "WAITING_TIME");
         public final static Property Hitchability = new Property(15, Integer.class, "Hitchability", false, "HITCHABILITY");
         public final static Property AttemptResult = new Property(16, Integer.class, "AttemptResult", false, "ATTEMPT_RESULT");
@@ -46,6 +46,7 @@ public class SpotDao extends AbstractDao<Spot, Long> {
         public final static Property Accuracy = new Property(21, Float.class, "Accuracy", false, "ACCURACY");
         public final static Property IsPartOfARoute = new Property(22, Boolean.class, "IsPartOfARoute", false, "IS_PART_OF_AROUTE");
         public final static Property IsHitchhikingSpot = new Property(23, Boolean.class, "IsHitchhikingSpot", false, "IS_HITCHHIKING_SPOT");
+        public final static Property IsNotHitchhikedFromHere = new Property(24, Boolean.class, "IsNotHitchhikedFromHere", false, "IS_GOT_OFF_HERE");
     }
 
 
@@ -84,7 +85,8 @@ public class SpotDao extends AbstractDao<Spot, Long> {
                 "\"HAS_ACCURACY\" INTEGER," + // 20: HasAccuracy
                 "\"ACCURACY\" REAL," + // 21: Accuracy
                 "\"IS_PART_OF_AROUTE\" INTEGER," + // 22: IsPartOfARoute
-                "\"IS_HITCHHIKING_SPOT\" INTEGER);"); // 23: IsHitchhikingSpot
+                "\"IS_HITCHHIKING_SPOT\" INTEGER," + // 23: IsHitchhikingSpot
+                "\"IS_GOT_OFF_HERE\" INTEGER);"); // 24: IsNotHitchhikedFromHere
     }
 
     /** Drops the underlying database table. */
@@ -162,9 +164,9 @@ public class SpotDao extends AbstractDao<Spot, Long> {
             stmt.bindString(13, Description);
         }
  
-        java.util.Date StartDateTime = entity.getStartDateTime();
+        Long StartDateTime = entity.getStartDateTimeMillis();
         if (StartDateTime != null) {
-            stmt.bindLong(14, StartDateTime.getTime());
+            stmt.bindLong(14, StartDateTime);
         }
  
         Integer WaitingTime = entity.getWaitingTime();
@@ -215,6 +217,11 @@ public class SpotDao extends AbstractDao<Spot, Long> {
         Boolean IsHitchhikingSpot = entity.getIsHitchhikingSpot();
         if (IsHitchhikingSpot != null) {
             stmt.bindLong(24, IsHitchhikingSpot ? 1L: 0L);
+        }
+ 
+        Boolean IsNotHitchhikedFromHere = entity.getIsNotHitchhikedFromHere();
+        if (IsNotHitchhikedFromHere != null) {
+            stmt.bindLong(25, IsNotHitchhikedFromHere ? 1L: 0L);
         }
     }
 
@@ -287,9 +294,9 @@ public class SpotDao extends AbstractDao<Spot, Long> {
             stmt.bindString(13, Description);
         }
  
-        java.util.Date StartDateTime = entity.getStartDateTime();
+        Long StartDateTime = entity.getStartDateTimeMillis();
         if (StartDateTime != null) {
-            stmt.bindLong(14, StartDateTime.getTime());
+            stmt.bindLong(14, StartDateTime);
         }
  
         Integer WaitingTime = entity.getWaitingTime();
@@ -341,6 +348,11 @@ public class SpotDao extends AbstractDao<Spot, Long> {
         if (IsHitchhikingSpot != null) {
             stmt.bindLong(24, IsHitchhikingSpot ? 1L: 0L);
         }
+ 
+        Boolean IsNotHitchhikedFromHere = entity.getIsNotHitchhikedFromHere();
+        if (IsNotHitchhikedFromHere != null) {
+            stmt.bindLong(25, IsNotHitchhikedFromHere ? 1L: 0L);
+        }
     }
 
     @Override
@@ -364,7 +376,7 @@ public class SpotDao extends AbstractDao<Spot, Long> {
             cursor.isNull(offset + 10) ? null : cursor.getShort(offset + 10) != 0, // IsReverseGeocoded
             cursor.isNull(offset + 11) ? null : cursor.getString(offset + 11), // Note
             cursor.isNull(offset + 12) ? null : cursor.getString(offset + 12), // Description
-            cursor.isNull(offset + 13) ? null : new java.util.Date(cursor.getLong(offset + 13)), // StartDateTime
+            cursor.isNull(offset + 13) ? null : cursor.getLong(offset + 13), // StartDateTime
             cursor.isNull(offset + 14) ? null : cursor.getInt(offset + 14), // WaitingTime
             cursor.isNull(offset + 15) ? null : cursor.getInt(offset + 15), // Hitchability
             cursor.isNull(offset + 16) ? null : cursor.getInt(offset + 16), // AttemptResult
@@ -374,7 +386,8 @@ public class SpotDao extends AbstractDao<Spot, Long> {
             cursor.isNull(offset + 20) ? null : cursor.getShort(offset + 20) != 0, // HasAccuracy
             cursor.isNull(offset + 21) ? null : cursor.getFloat(offset + 21), // Accuracy
             cursor.isNull(offset + 22) ? null : cursor.getShort(offset + 22) != 0, // IsPartOfARoute
-            cursor.isNull(offset + 23) ? null : cursor.getShort(offset + 23) != 0 // IsHitchhikingSpot
+            cursor.isNull(offset + 23) ? null : cursor.getShort(offset + 23) != 0, // IsHitchhikingSpot
+            cursor.isNull(offset + 24) ? null : cursor.getShort(offset + 24) != 0 // IsNotHitchhikedFromHere
         );
         return entity;
     }
@@ -394,7 +407,7 @@ public class SpotDao extends AbstractDao<Spot, Long> {
         entity.setIsReverseGeocoded(cursor.isNull(offset + 10) ? null : cursor.getShort(offset + 10) != 0);
         entity.setNote(cursor.isNull(offset + 11) ? null : cursor.getString(offset + 11));
         entity.setDescription(cursor.isNull(offset + 12) ? null : cursor.getString(offset + 12));
-        entity.setStartDateTime(cursor.isNull(offset + 13) ? null : new java.util.Date(cursor.getLong(offset + 13)));
+        entity.setStartDateTimeMillis(cursor.isNull(offset + 13) ? null : cursor.getLong(offset + 13));
         entity.setWaitingTime(cursor.isNull(offset + 14) ? null : cursor.getInt(offset + 14));
         entity.setHitchability(cursor.isNull(offset + 15) ? null : cursor.getInt(offset + 15));
         entity.setAttemptResult(cursor.isNull(offset + 16) ? null : cursor.getInt(offset + 16));
@@ -405,6 +418,7 @@ public class SpotDao extends AbstractDao<Spot, Long> {
         entity.setAccuracy(cursor.isNull(offset + 21) ? null : cursor.getFloat(offset + 21));
         entity.setIsPartOfARoute(cursor.isNull(offset + 22) ? null : cursor.getShort(offset + 22) != 0);
         entity.setIsHitchhikingSpot(cursor.isNull(offset + 23) ? null : cursor.getShort(offset + 23) != 0);
+        entity.setIsNotHitchhikedFromHere(cursor.isNull(offset + 24) ? null : cursor.getShort(offset + 24) != 0);
      }
     
     @Override
