@@ -1,7 +1,6 @@
 package com.myhitchhikingspots.utilities;
 
 import java.io.File;
-import java.util.ArrayList;
 
 import android.app.Activity;
 import android.content.Context;
@@ -14,7 +13,7 @@ public class FilePickerDialog {
 
     public void openDialog(final Activity activity, final Context context) {
         File mPath = new File(Constants.EXPORTED_DB_STORAGE_PATH);
-        FileDialog fileDialog = new FileDialog(activity, mPath, ".csv", ".db");
+        FileDialog fileDialog = new FileDialog(activity, mPath, Constants.EXPORT_DB_AS_CSV_FILE_EXTENSION, Constants.INTERNAL_DB_FILE_EXTENSION);
         fileDialog.addFileListener(new FileDialog.FileSelectedListener() {
 
             @Override
